Add Kernel.org Builders
Two new builders have been added which build against the latest
Linux kernel.org source.  This ensures we detect build issues as
soon as those changes are merged upstream.

* Kernel.org Default x86_64 - Builds ZFS in-tree against the
  latest kernel with a default configuration, `make defconfig`.
  The following additional options are added:

    CONFIG_CRYPTO_PCOMP=y
    CONFIG_CRYPTO_ZLIB=y
    CONFIG_ZLIB_DEFLATE=y
    CONFIG_KALLSYMS=y
    CONFIG_EFI_PARTITION=y

* Kernel.org Built-in x86_64 - Adds ZFS to the kernel source
  and builds it as part of the normal kernel build.  This is
  done to regularly excercise the `--enable-linux-builtin`
  functionality.

    CONFIG_SPL=y
    CONFIG_ZFS=y

This commit has the additional benefit of laying the ground work
so a Lustre build step may be added for specific builders.  Having
a Lustre build would act as a cross-check to ensure the ABI remains
stable.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>