CentOS 6 Bootstrapping
Because the buildbot-slave package is not readily available for
CentOS 6 from the major repositories bootstrapping relied on
easy_install.  Recent experience has shown this to be fragile
under CentOS 6 and prone to problems.  The most recent being a
python 2.7 dependency which is not provided by CentOS 6.  To
avoid the issue, and speed up bootstrapping, install a prebuilt
version of buildbot-slave.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>