Boot TEST builders on to latest kernel
As part of bootstrapping a TEST builder the latest distribution
supplied kernel should be installed and rebooted on to before
launching tests.  This ensures we're always running the testing
using the latest distribution provided kernel.  In addition,
it prevents testing failures due to the running kernel being
out of sync with the installed kernel development packages.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>