Treat a non-zero ztest exit code as a failure
The ztest stress test can return multiple non-zero exit codes to
indicate how the test failed.  The actual exit code needs to be
logged and a 1 (FAILURE) returned to ensure the buildbot correctly
categorizes the test result.  For example, an exit code of 3 is
used for a skipped test which would suppress the failure.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>