Optimize how slaves are assigned builds.
It is preferable to assign a build to a slave who is
idle (already spun up and waiting) over a slave that has yet
to be started or a busy one. If there are no idle slaves,
and there is one that can be spun up, then spin one up.
Otherwise, randomly choose among our busy slaves.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>