Force root volumes to be deleted upon termination on EC2 volumes.
This patch is for buildbot. It introduces the ability to set all root
volumes of an instance to be deleted upon termination. This change
does not impact additionally attached volumes.

Spot instances will make root volumes and are similar to on demand
instances.  They will not destroy a volume unless delete on termination
is set.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Closes #20