Decrease build_wait_timeout to 30 minutes
Now that we're able to start the buildslaves in parallel there is
little benefit to leaving them idle for very long.  Assume the VM
was booted 3 builds or less each of which took less than 10 minutes
(they average about 6-7 minutes).  Then wait for the majority of
the remaining billable hour just in case a new job comes in.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>