Skip SPLAT tests by default
Since the SPL is updated far less frequently than the ZFS code
there's no reason to run the SPLAT tests for every commit.
They only need to be run when a new kernel is released or when
there are changes to the SPL itself.  To enable these tests
in a PR add the following line to you get commit messages.

TEST_SPLAT_SKIP="No"

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>