Introduce a CentOS 7 Release BUILD builder
In order to ensure that we do not break lustre builds,
introduce a new CentOS 7 builder that will build zfs without
debug enabled and build lustre. The current stable branch is
for lustre 2.10 and it's called b2_10, so we will build
that to ensure we don't break the current stable release.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>