bb-test-ztest.sh: Trap SIGTERM
If zloop.sh hangs and exceeds the allowed buildbot log timeout it
will be sent SIGTERM.   In this case the EXIT handler will not be
run and the modules loaded.  Subsequent test build steps will fail
when trying to load the modules.

Additionally remove the redundant console dump.  This test runs
entirely in user space and it's unlikely to trigger a problem in
the loaded modules.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>