Disable zfsstress by default on TEST builders
Now that the vast majority of the ZTS is running there is less
value in always running this test suite.  It can be enabled when
needed by setting TEST_ZFSSTRESS_SKIP="No" in your commit message.

As an aside, these test scripts predate getting the ZTS framework
working on Linux.  It would be useful to adapt them to the ZTS and
include them under tests/zfs-tests/tests/stress/.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>