Fix ZTS Amazon 2 failures
The Amazon Linux 2 AMI includes a udev rule which creates symlinks
from /dev/sda to /dev/xvda in order to "maintain a consistent naming
scheme".  This conflicts with the names assigned to real scsi_debug
device resulting in test failures.  Therefore, the rule which creates
these aliases must be removed when bootstrapping the instance for
ZTS testing.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>