Correct built in kernel builds
spl and zfs were not included in built in kernel builds
since the build_spl and build_zfs properties were preventing
the repositories from being cloned. Correct the BUILD build
factory to account for built in builds and only perform the
build process for spl and zfs if the current build is not
a built in build.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>