Replace Kernel.org builder with Fedora Rawhide
Use the Fedora Rawhide release to perform testing on the latest
kernels.  By using these kernels which are built nightly we can
easily run the ZFS Test Suite against the latest kernels.

Additionally, this change ensures we're aware of and have time
to address and changes in the Fedora packaging guidelines before
the next version is officially released.

For the moment, non-debug packages must be built until all new
warnings from gcc 8 are resolved and the non-debug kernel is used.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>