Fix Fedora Rawhide bootstrapping
Fedora 29 will ship with buildbot v1.0 which is incompatible with
v0.8 which is used on the buildbot master.  Therefore, install the
buildbot-slave packages from pip which still provides the older
compatible verison.

This is at best a stop gap measure until we can move the buildbot
master forward to a newer version of buildbot.  Newer versions of
the master are apparently compatible with older slave versions.

Additionally update the Fedora Rawhide AMI.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>