Ensure EPEL is available for CentOS/RHEL builders
When using the generic AMIs, epel-release is not installed by default,
thus the EPEL repo isn't even available, which causes part of the dependency
installation to fail.

To resolve this, we'll install epel-release and ensure the epel repo is disabled,
so that we only use EPEL when we need to.

Signed-off-by: Neal Gompa <ngompa@datto.com>