Add make build-dep for Fedora, RHEL, and SUSE
When using the official Fedora Cloud base images, we wind up setting up
the build environment without having make installed, which causes the
configure script to fail when it tries to do compilation-based tests.

It also appeared that this wasn't properly set for RHEL and SUSE, so
that is fixed now, too.

Signed-off-by: Neal Gompa <ngompa@datto.com>