Add support for VPC and storage configuration for build slaves
For hybrid cloud deployments where part of the resources are private
or partitioned, it is necessary for the build slaves to be able to be
configured to be instantiated with a VPC configured.

In addition, for using official cloud images from CentOS and Fedora
(and potentially any other distribution's official images), buildbot
needs to provision the build slave in EC2 with the ephemeral virtual
disks rather than having them manually configured by cloning an AMI
in the AWS console.

Finally, a trivial change to backfill in the description for one of
the patches is included in this change, as well as a note of what
buildbot release the entire patch set will apply on top of.

Signed-off-by: Neal Gompa <ngompa@datto.com>