Introduce an override for kernel built
The Kernel Builders often fail to build because
new patches have been introduced in the upstream linux
kernel. The builders will often fail builds for quite
some time which defeats the purpose. Introduce a method
to override which tag we want to build while build issues
from the linux kernel's master branch are resolved.

The tag can be specified in the `kernel.tag` file within
the build master's working directory.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>