Add back dropped default setting for price_multiplier
Prior to the remaining production patches being committed in
3898450b79640d23df62cd5e2851c0f1e5a7232c, the price_multiplier setting
did not work and caused it to crash on spinning up slaves. Now that
the patches are published, the setting is confirmed to work again.

In addition, a trivial change to re-order the patches relatively based
on when they were created to ensure that they can be applied incrementally
is included along with updated patch documentation.

Signed-off-by: Neal Gompa <ngompa@datto.com>