Conditionally checkout and build the spl (#162)
Update the CustomGitHubEventHandler to conditionally checkout and
build the SPL when required.  This is currently possible using
the "buildspl" property, but that propery is set by the builder
and cannot by overidden by a change.

Therefore, the "buildspl" property has been switched to a "Change"
properly so it can be set correctly in CustomGitHubEventHandler.
The "buildspl" properly will be set to "yes" when building any
commit on the zfs-0.7-release branch.  Otherwise, it will be set
to "no" and the spl checkout and build will be skipped.

As part of this change the CustomGitHubEventHandler() was moved
in to its own file to simplify the master.cfg.  Some additional
minor cleanup was also included.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>