Preserve test results for analysis
Often it is useful to be able to download the artifacts from the
CI system run for futher analysys.  Support this by updating the
test scripts to place their build artifacts and test logs in a
known location so they can be uploaded to the build master.

The test results are merged together in to a single directory
using the git commit description as the name.  The log archives
for each builder are then placed in subdirectores.  In order to
prevent file name collisions both the commit hash and a time
stamp are includes in the archive file names when possible.

This change additionally relocates the performance builder to
an AWS d2.xlarge instance.  This ensures a pristine consistent
environment for performance testing.  One potential concern is
an increased variance in test runs since each run is given a
unique instance, but thus far this hasn't been a problem.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>