Add python3-devel packages
On distributions which provide Python 3 packages, install the
python3-dev, python3-setuptools, and python3-cffi packages.
They are required to build the pythonX-pyzfs package.  The
Python 2 versions of these packages have been left installed
as a testing convenience.

Update the Amazon 2 AMI which was getting a little stale.

Remove flake8 install on Amazon which was always failing
since pip is not available.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>