Only test the top commit for large PRs
To prevent the buildbot from being overwhelmed by PRs and
pushes containing a large number of individual commits impose
a limit.

For PRs containing more than 5 commits only test the top commit.
All other commits in the PR stack are ignored.

For pushes only test the first 10 commits.  A larger limit is
allowed for pushes since they are less frequent and only occur
when a PR is merged to a branch by its maintainer.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>