Disable coverage builder by default for PRs
Only run the coverage builder for a PR when requested to do.  The
builder can be requested by adding the `Requires-builders:` keyword
and a list of builders to the git commit message.  For example, the
following line will request all current builders:

Requires-builders: arch,style,amazon2,centos7,centos8,debian10,
    fedora32,ubuntu18,ubuntu20,builtin,freebsd12,freebsd13,coverage

The coverage build is still run on all merges to master.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>