FreeBSD: Don't add /mnt to fstab
This doesn't really affect the CI, but when following these scripts to
manually prepare an EC2 instance, having the instance storage in fstab
makes instances unbootable after being stopped and started again.  The
instance storage is ephemeral, so the mount fails due to the new
instance not having the partition, and the system drops to single user
mode.  With no console access, the volume is effectively bricked.

Just mount the partition without adding it to fstab.