FreeBSD: Split up test dependencies
Occasionally packages are broken in the FreeBSD head repo.  When one
package fails to install, the rest of the packages in the same command
invocation fail as well.

Split up the test dependencies into groups by necessity, so that when
a non-essential utility such as gdb fails to install, we still install
essential packages like ksh93 and python.  This allows most tests to
run as long as we aren't missing something essential.

Signed-off-by: Ryan Moeller <ryan@iXsystems.com>