Remove Amazon Linux 2 builders
Further simplify the testing matrix by retiring the Amazon Linux
builders in order to focus on the other major platforms/distributions.

For the moment support for the Amazon environment hasn't been removed
from the scripts.

Note this case also added bison and flex to the Ubuntu environment
to allow it to be used as the builtin builder.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>