Use Almalinux 8 AMI in place of CentOS 8
After RedHat reported that CentOS would prematurely be EOL'd on
Dec 31 2021, we have decided to switch to Almalinux for CentOS 8
builders.  Almalinux is a open source, supported, clone on RHEL/CentOS 8.
We still use the "CentOS" naming in buildbot since it's still CentOS
compatible, and so we don't have to update all our bash scripts.

Signed-off-by: Tony Hutter <hutter2@llnl.gov>