Revert "bb-bootstrap, install a newer kernel on buster arm64"
This reverts commit 73521285 which was originally added to work around
some 'limitations' with the default Buster kernel on arm64.  It's been
over a year since this change was made and the default kernel seems
to work fine now.  Furthermore, we encountered an issue with mismatched
packages in the linux-image-cloud-arm64 repository.  So we're going to
try switching back to the default kernel.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>