FreeBSD: Use public SSM parameters for latest AMIs
FreeBSD now registers the latest AMI IDs as public parameters in the AWS
SSM parameter store.  EC2 APIs accept SSM parameters in place of an AMI
and will resolve the AMI from the parameter value.  Use this to replace
the current hacked in method of determining the latest FreeBSD AMIs at
run time.

More info on the SSM parameters:
https://www.daemonology.net/blog/2021-08-31-FreeBSD-AMI-SSM-Public-Parameters.html