FreeBSD: bootstrap: Handle an existing swap
Recently a script built into FreeBSD AMIs that sets up a swap on
ephemeral storage was updated to work for NVMe storage as well.  This
interferes with our attempt to partition and format the device for /mnt.

As a result, nothing gets mounted on /mnt, so the tests fill up the root
filesystem and evenually fail.

Remedy this by checking which partition was added rather than assuming
it was the first one.